refactor: Hashtag Model 구조 및 Resolver 파라미터 수정
