+++ src/domains/hashtags/models/index.js
@@ -1,5 +1,7 @@
 import {hashtagTypeDefs} from './@types/index.js';
+import Hashtag from './hashtag.js';
 
 export {
   hashtagTypeDefs,
+  Hashtag,
 };
