+++ src/domains/hashtags/resolvers/hashtags.js
@@ -1,12 +1,7 @@
-import Hashtag from '../models/schema/index.js';
+import {Hashtag} from '../models/index.js';
 
-const getHashtags = async () => {
-  try {
-    return await Hashtag.find();
-  } catch (err) {
-    console.log(err);
-    throw err;
-  }
+const getHashtags = async (parent, args, context, info) => {
+  return await Hashtag.find();
 };
 
 export default getHashtags;
